% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_structure_replicate.R
\name{folder_structure_replicate}
\alias{folder_structure_replicate}
\title{Replicate the folder structure of a given directory}
\usage{
folder_structure_replicate(dir, to)
}
\arguments{
\item{dir}{Character 1L. Path of directory which structure will be replicated}

\item{to}{Character 1L. Path of an output directory in which replicated structured will be placed}
}
\value{
No return value, called for side effects. The directory
structure of \code{dir} is recreated inside the \code{to} directory.
}
\description{
Replicate the folder structure of a given directory
}
\examples{
library(magrittr)

temp_dir_to_replicate <- tempfile()
dir.create(temp_dir_to_replicate)

dir.create(file.path(temp_dir_to_replicate, "dir1"))
dir.create(file.path(temp_dir_to_replicate, "dir2"))

temp_dir_out <- tempfile()
dir.create(temp_dir_out)

folder_structure_replicate(
  dir = temp_dir_to_replicate, 
  to = temp_dir_out)

unlink(temp_dir_to_replicate)
unlink(temp_dir_out)
}
