% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industtry-package.R
\docType{package}
\name{industtry-package}
\alias{industtry}
\alias{industtry-package}
\title{industtry: A Toolkit for Structured Datasets Exploitation}
\description{
A set of functions for exploiting collections of structured datasets (data frames) at scale. Core capabilities include batch inspection and profiling, variable presence detection across datasets, type consistency checking, format conversion, and file management helpers. Designed for workflows involving multiple related datasets with a common structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danielrak/industtry}
  \item Report bugs at \url{https://github.com/danielrak/industtry/issues}
}

}
\author{
\strong{Maintainer}: Daniel Rakotomalala \email{rakdanielh@gmail.com}

}
\keyword{internal}
