% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_convert_r.R
\name{mask_convert_r}
\alias{mask_convert_r}
\title{Create an excel mask compatible with the convert_r() function}
\usage{
mask_convert_r(output_path, output_filename = "mask_convert_r.xlsx")
}
\arguments{
\item{output_path}{Character 1L. Folder path where the mask will be created}

\item{output_filename}{Character 1L. File name (with extension) of the mask}
}
\value{
No return value, called for side effects. An Excel file
(\code{.xlsx}) is written to \code{file.path(output_path, output_filename)}
containing an empty template with columns \code{folder_path}, \code{file},
\code{converted_file}, and \code{to_convert}.
}
\description{
Create an excel mask compatible with the convert_r() function
}
\examples{
mydir <- file.path(tempdir(), "convert_r_tests_examples")
dir.create(mydir)

mask_convert_r(output_path = mydir)

list.files(mydir)

unlink(mydir, recursive = TRUE)
}
