% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_rename_r.R
\name{mask_rename_r}
\alias{mask_rename_r}
\title{Create an excel mask compatible with the rename_r() function.}
\usage{
mask_rename_r(input_path, output_filename = "mask_rename_r.xlsx")
}
\arguments{
\item{input_path}{Character 1L. Folder containing the set of files to rename}

\item{output_filename}{Character 1L. File name of the excel mask.}
}
\value{
No return value, called for side effects. An Excel mask file
is written to \code{input_path}.
}
\description{
Create an excel mask compatible with the rename_r() function. This must be used on a collection of files, i.e stored within the same folder.
}
\examples{
library(magrittr)
data(cars)
data(mtcars)

mydir <- tempfile()
dir.create(mydir)

saveRDS(cars, file.path(mydir, "cars.rds"))
saveRDS(mtcars, file.path(mydir, "mtcars.rds"))

list.files(mydir)

mask_rename_r(input_path = mydir)

list.files(mydir)

readxl::read_xlsx(file.path(mydir, "mask_rename_r.xlsx"))

unlink(mydir, recursive = TRUE)
}
