% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_multiple.R
\name{replace_multiple}
\alias{replace_multiple}
\title{Replace character vector values using a correspondence approach}
\usage{
replace_multiple(input_vector, replacements, replace_all = FALSE)
}
\arguments{
\item{input_vector}{Character. Character vector on which replacements take place.}

\item{replacements}{Character. Named character vector defining replacement
correspondences (names = patterns, values = replacements).}

\item{replace_all}{Logical. If \code{TRUE}, \code{stringr::str_replace_all()}
is used instead of \code{stringr::str_replace()}.}
}
\value{
Character vector with replacements applied.
}
\description{
Replace character vector values using a correspondence approach
}
\examples{
input <- c("one-one", "two-two-one", "three-three-two")

replace_multiple(input,
                 replacements =
                   c("one" = "1", "two" = "2",
                     "three" = "3"))

replace_multiple(input,
                 replacements =
                   c("one" = "1", "two" = "2",
                     "three" = "3"),
                 replace_all = TRUE)
}
