% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrutr-package.R
\docType{package}
\name{scrutr-package}
\alias{scrutr}
\alias{scrutr-package}
\title{scrutr: Scrutinizing Collections of Structured Datasets}
\description{
Provides a coherent interface for exploring and transforming multiple related data frames that share a common structure. Complements single-dataset inspection tools by operating across an entire collection at once. Also includes lightweight utilities for related file and folder management tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danielrak/scrutr}
  \item Report bugs at \url{https://github.com/danielrak/scrutr/issues}
}

}
\author{
\strong{Maintainer}: Daniel Rakotomalala \email{rakdanielh@gmail.com}

}
\keyword{internal}
