% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_detect_not_everywhere.R
\name{vars_detect_not_everywhere}
\alias{vars_detect_not_everywhere}
\title{Variable detection - inconsistent patterns}
\usage{
vars_detect_not_everywhere(vars_detect_table)
}
\arguments{
\item{vars_detect_table}{Output of the \code{vars_detect()} function.}
}
\value{
A subset of \code{vars_detect_table} containing only variables
that are not present in every dataset.
}
\description{
Variable detection - inconsistent patterns
}
\examples{
data_list <- list(cars = cars, mtcars = mtcars)
vdetect_table <- vars_detect(data_list)
vars_detect_not_everywhere(vdetect_table)
}
